Session 6: INT8 activation quant hypothesis falsified
Session 5 corrected PCC attribution to "bf16 drift". Session 6 tested
whether implementing per-token INT8 absmax ActQuant inside ternary_matmul
(matching bitnet.cpp / HF AutoBitLinear) would close the gap.

Measurement: toggle ActQuant ON/OFF in HF bf16 reference, measure TT PCC
against each.

  HF ActQuant ON  (default)   TT vs HF PCC = 0.981319
  HF ActQuant OFF (identity)  TT vs HF PCC = 0.980143

Delta = +0.0012, noise. Reproducing the same quant step inside the
TT kernel can at best close 0.001, not the 0.019 gap. US-602 / US-603
kernel work closed as obsolete before any tt-metal edit.

New references benchmarked:
  TT vs HF fp32            PCC 0.980679
  TT vs HF bf16 (+ActQuant) PCC 0.981319
  TT vs HF bf16 (-ActQuant) PCC 0.980143
  TT vs bitnet.cpp (INT8×INT2 CPU) PCC 0.967860

Real ~0.019 gap source: op-implementation differences between tt-metal
bf16 ops and PyTorch/CUDA bf16 (SDPA reductions, RMSNorm ordering, RoPE
fused kernels, residual add order). Not addressable by BitNet-specific
quantisation changes.

Additions:
- bench_vs_bitnetcpp.py, compare TT logits vs bitnet.cpp binary dump
- bench_vs_hf_noquant.py, HF bf16 reference with ActQuant disabled
- docs/session_6_int8_act.md, full measurement + decision record
- generator_batch32.py tokens cast int64 -> int32 (torch 2.2.2 lacks
  uint32, the int64->ttnn.uint32 path fails after recent deps pull)

No kernel changes. Current HEAD holds 74.21 t/s / PCC 0.981 vs HF bf16.

+++ src/bitnet_tt/inference/generator_batch32.py
@@ -1254,7 +1254,7 @@ class Batch32Generator:
         batch_size, seq_len = tokens.shape
 
         tokens_tt = ttnn.from_torch(
-            torch.from_numpy(tokens.astype(np.int64)),
+            torch.from_numpy(tokens.astype(np.int32)),
             dtype=ttnn.uint32,
             layout=ttnn.ROW_MAJOR_LAYOUT,
             device=self.device,
